🔧 fix: Server action consolidates and organizes annotations
